+++ test/inspec/author-publish-dispatcher_spec.rb
@@ -160,6 +160,42 @@ apache_https_port = @hiera.lookup('aem_curator::install_dispatcher::apache_https
 
 cert_filename = @hiera.lookup('aem_curator::install_dispatcher::cert_filename', nil, @scope)
 
+dispatcher_id = @hiera.lookup('aem_curator::install_dispatcher::aem_id', 'dispatcher', @scope)
+
+setup_data_volume = @hiera.lookup('aem_curator::install_dispatcher::setup_data_volume', nil, @scope)
+
+# data_volume_device = @hiera.lookup('aem_curator::install_dispatcher::data_volume_device', nil, @scope)
+data_volume_device = '/dev/xvdd'
+
+# data_volume_mount_point = @hiera.lookup('aem_curator::install_dispatcher::data_volume_mount_point', nil, @scope)
+data_volume_mount_point = '/mnt/ebs3'
+
+docroot_dir = @hiera.lookup('aem_curator::install_dispatcher::docroot_dir', '/var/www/html', @scope)
+
+apache_user = @hiera.lookup('aem_curator::install_dispatcher::apache_user', 'apache', @scope)
+
+apache_group = @hiera.lookup('aem_curator::install_dispatcher::apache_group', 'apache', @scope)
+
+if setup_data_volume
+  describe mount(data_volume_mount_point) do
+    it { should be_mounted }
+    its('device') { should eq data_volume_device }
+  end
+
+  describe file(docroot_dir) do
+    it { should exist }
+    it { should be_symlink }
+    its('link_path') { should eq "#{data_volume_mount_point}/#{dispatcher_id}" }
+  end
+else
+  describe file(docroot_dir) do
+    it { should exist }
+    it { should be_directory }
+    its('owner') { should eq apache_user }
+    its('group') { should eq apache_group }
+  end
+end
+
 describe package(apache_package) do
   it { should be_installed }
 end