+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added 
+### Added
+- Add InSpec tests for checking Dispatcher Data Volume
 - Add python3 and python2.7 virtualenv
 - Add snapshot_users in order to allow destination accounts to copy shared AMIs